Fixed 'value was either too large or too small for an int32' error on deserialization'

+++ SharpBucket/V2/Pocos/IteratorBasedPage.cs
@@ -9,6 +9,6 @@ namespace SharpBucket.V2.Pocos {
         public int? pagelen { get; set; }
         public string next { get; set; }
         public List<T> values { get; set; }
-        public int? size { get; set; }
+        public ulong? size { get; set; }
     }
 }
